requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
flatbuffers==24.3.25
fonttools==4.52.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jax==0.4.28
jaxlib==0.4.28
kiwisolver==1.4.5
matplotlib==3.9.0
mediapipe==0.10.14
ml-dtypes==0.4.0
MouseInfo==0.1.3
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
protobuf==4.25.3
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyGetWindow==0.0.9
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pywin32-ctypes==0.2.2
requests==2.32.3
requests-oauthlib==1.3.1
schedule==1.2.2
scipy==1.13.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
tqdm==4.66.4
tweepy==4.14.0
typing_extensions==4.12.2
urllib3==2.2.2

